loading model test fixes
